requirements.txt
sounddevice
numpy
scipy
openai
python-dotenv